requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
imbalanced-learn>=0.11.0

# # Feature Engineering
# category_encoders>=2.6.0
# feature_engine>=1.5.2

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
joblib>=1.2.0
tqdm>=4.65.0
PyYAML>=6.0

# Web Application
streamlit>=1.28.0
streamlit-metrics>=0.1.0
streamlit-aggrid>=0.3.0